Introduce Travis CI configuration file
With this commit Travis continuous integration system is configured.
This in turn allows us to build our project for GHC 7.8 and 7.6 in clean
slate VM instances with every push to GitHub. It does a better job than
our private Jenkins installation and it is free so there is no need not
to use it.