Evaluation framework.
eval/infer.sh is the main program to use
eval/files.sh identifies all of the files to be tested
eval/is_valid_dir.sh kicks out anything with preprocessing
eval/inferRaw.sh is sometimes helpful for debugging to see the raw output

The corpus is assumed to be in $HOME/src/corpus but you can edit eval/files.sh
to change that if you like. eval/infer.sh assumes that camfort is in
dist/build/camfort where cabal puts it. You can change that if you like.