Update to use GHC 8.0.1
Some of the implicit extensions are made explicit as it is now enforced
by the compiler. In particular FlexibleContexts and FlexibleInstances.

Monad instances is now derived for tuple in the base, thus the instance
in Traverse.hs is deleted.

New stack manifest is generated to use a nightly build as stable
releases do not yet use GHC 8.