Introduce concept of "ModFile" with extension ".camfortmod"
ModFile currently stores two pieces of information about the modules in a file:
the module map for the renamer, and the template map for the units solver.
Then combinedModuleMap can be used to bring all of the module maps together.
And the Renamer has a function called analyseRenamesWithModuleMap that can
accept a module map as a parameter as a starting point for renames in a
multi-module context.