Subtle condition for instantiating explicitly annotated parametric polymorphic unit variables.
At the top-level of a call-stack we need to preserve the abstract EAPs and use
Gaussian elimination to solve for them. But in nested calls the EAPs must be
converted into concrete uses so that each use gets its own version. This is
tantamount to the rule for polymorphic lambda application.