A first stab at implicit polymorphism inference.
There seems to be a contradiction between the interpretation of generated
implicit polymorphic unit variables (represented as UnitParamPosAbs here) and
the consistency of the corresponding set of linear equations. The equations do
solve consistently if you put the generated polymorphic variables on the 'left
hand side' but aren't interpreted correctly by the transformation into
assignments. So, I have changed the code so that both interpretations are used:
first, the variables are treated as 'LHS' columns for the purpose of solving,
and then, they are treated as 'RHS' for the purpose of interpreting the results
of solving. Weird, but so far it works. For now.

Still need to cook up a system of converting UnitParamPosAbs values into
something more human readable such as "'a".